Move publish plugin to convention plugin
