+++ kt/build-logic/convention/src/main/kotlin/publish/PublishToMavenCentralPlugin.kt
@@ -1,4 +1,4 @@
-package godot.publish.gradle.mavencentral
+package publish
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
