--- kt/plugins/godot-publish-gradle-plugin/build.gradle.kts
@@ -1,20 +1,0 @@
-
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-    `java-gradle-plugin`
-}
-
-kotlin {
-    jvmToolchain(17)
-}
-
-gradlePlugin {
-    plugins {
-        create("godotPublishPlugin") {
-            id = "com.utopia-rise.godot-publish"
-            displayName = "Gradle plugin for publishing godot kotlin jvm to maven central"
-            implementationClass = "godot.publish.gradle.mavencentral.PublishToMavenCentralPlugin"
-        }
-    }
-    isAutomatedPublishing = false
-}
