--- kt/plugins/godot-publish-gradle-plugin/settings.gradle.kts
@@ -1,29 +1,0 @@
-
-pluginManagement {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        gradlePluginPortal()
-        google()
-    }
-}
-
-dependencyResolutionManagement {
-    @Suppress("UnstableApiUsage")
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        google()
-    }
-
-    versionCatalogs {
-        create("libs") {
-            from(files("../../gradle/libs.versions.toml"))
-        }
-    }
-}
-
-plugins {
-    // to automatically download the toolchain jdk if missing
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.6.0"
-}
