+++ kt/settings.gradle.kts
@@ -12,7 +12,6 @@ pluginManagement {
     resolutionStrategy.eachPlugin {
         when(requested.id.id) {
             "com.utopia-rise.api-generator" -> useModule("com.utopia-rise:api-generator:0.0.1")
-            "com.utopia-rise.godot-publish" -> useModule("com.utopia-rise:godot-publish-gradle-plugin:0.0.1")
             "com.github.johnrengelman.shadow" -> useVersion("7.1.2")
         }
     }
@@ -53,12 +52,6 @@ subdir("plugins") {
     include("godot-gradle-plugin")
     include("godot-intellij-plugin")
     include("godot-plugins-common")
-
-    this.includeBuild("godot-publish-gradle-plugin") {
-        dependencySubstitution {
-            substitute(module("com.utopia-rise:godot-publish-gradle-plugin")).using(project(":"))
-        }
-    }
 }
 
 subdir("utils") {
